requirements.txt
Flask==0.9
Flask-RESTful==0.2.1
RPi.GPIO==0.5.2a